requirements.txt
streamlit==1.36.0
openai==1.35.7
requests==2.28.2
langchain==0.2.5
langchain-community==0.2.5
langchain-openai==0.1.14
python-dotenv==1.0.1
